p29.30 refactor: thread ids.SplitID everywhere
Convert the split entity id from bare int64 to the distinct ids.SplitID type
across every layer: sqlc models (splits.ID, SplitsVersion.EntityID) + the split
queries (InsertSplit return, DeleteSplit/GetSplitForReconcile params,
UpdateSplitParams.ID, RepointSplitAccountParams.ID, InsertSplitVersionParams.ID_2,
SetSplitReconciliationParams.ID, SplitIdsByAccount/FinalizedReconciledSplitIDs
[]ids.SplitID returns, the split-id-carrying rows: SplitVersionHistory/AsOf
EntityID, SplitsByAccountCurrency/Workspace/ClearedSplits/GetSplitForReconcile ID,
Drill/Register/FundLedger SplitID); the store seam (the editor's SplitInput.ID /
resolvedSplit.id *ids.SplitID + the split-id-keyed UpdateTransaction maps
liveByID/liveAccountByID/lockedSet/inputIDs/kept/seenID, insertSplitVersion +
validateAndResolve, SplitState/HistSplitState/SplitDiff/AccountSplitRef ID,
TransactionAsOf's seen map + history reconstructStates live/prevSplit maps,
SetSplitReconciled + the recon store-facing SplitID fields, RegisterCursor.SplitID,
SplitIDsForAccount); synth (the ExtendReconciliation toClear list); the web layer
(the editor split-id round-trip: parse ids.SplitID(parseID(s)) at the hidden-form
boundary, render via int64(sp.ID); the reconcile toggle parse + comparison; the
register cursor + counter maps; drill's resolveDrillCounter selfSplitID); and the
full test fan-out.

Purely types, zero behavior change: goldens byte-identical, make check clean.
Boundary discipline (rule 15) — splits_versions.ID snapshot PKs + every AccountID
on these rows stay int64 (AccountID is the next commit); the editor id round-trip
uses *ids.SplitID (a Go pointer, NOT the sql.NullInt64 ids.Ptr helper); render-only
web fields + itoa/FormatInt/reconToggleID/reconRowID string-id helpers keep int64
and cast int64(x); driver.Value binds in tests cast int64(x). ST1019: files that
shadow `ids` with a local var reuse the existing entids alias (synth seams/fixture)
or a renamed local (sids/txnIDs in balances/accounts).

+++ internal/web/register_test.go
@@ -498,9 +498,9 @@ func TestRegisterParentRollupCounterAccount(t *testing.T) {
 	for _, sp := range splits {
 		switch sp.AccountID {
 		case boa:
-			boaSplit = sp.ID
+			boaSplit = int64(sp.ID)
 		case wf:
-			wfSplit = sp.ID
+			wfSplit = int64(sp.ID)
 		}
 	}
 
@@ -694,7 +694,7 @@ func TestRegisterHtmxPaging(t *testing.T) {
 	// and NOT re-render the first row (append semantics).
 	q := url.Values{}
 	q.Set("cursor_date", cur.Date)
-	q.Set("cursor_id", strconv.FormatInt(cur.SplitID, 10))
+	q.Set("cursor_id", strconv.FormatInt(int64(cur.SplitID), 10))
 	frag := asUser(t, e.h, e.sm, reader, http.MethodGet, base+"?"+q.Encode(), nil)
 	if frag.Code != http.StatusOK {
 		t.Fatalf("htmx paging: %d body=%s", frag.Code, frag.Body.String())
